testtask/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
black==24.10.0
bleach==6.2.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.5
django-environ==0.12.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
flake8==7.1.1
frozenlist==1.5.0
geoip2==4.8.1
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
idna==3.10
isort==5.13.2
kombu==5.4.2
Markdown==3.7
maxminddb==2.6.3
mccabe==0.7.0
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.18.3
